Fixed an issue where Muzei couldn't display default artwork
For some reaosn the Muzei couldn't resolve the uri returned from the media picker intent.. something to do with Google's photo picker and possibly weird permission issues.

The workaround here is to store the default wallpaper locally in the app's external files dir. This isn't a bad idea anyway, it means the default artwork will always be available, even if the source is deleted.